refactor: 미들웨어 위치 변경
